+++ CHANGES.md
@@ -6,6 +6,13 @@ See [compatibility docs](docs/compatibility.md) for details about backward compa
 
 ## CHANGE LOG ##
 
+* @dev
+   * **BACKWARD INCOMPATIBLE CHANGES** ([more info](docs/compatibility.md))
+   * Reworked `ExceptionHandlerHelper` configuration. Now, you will be able to easily configure every
+     HttpException for each HTTP status code you want. Separate `ExceptionHandler::TYPE_HTTP_NOT_FOUND_KEY`
+     and all related stuff, incl. localization key `http_not_found`, configuration is now replace with more
+     flexible generic code that provides error messages for all supported HTTP codes from in range `400-599`.
+
 * 6.3.1 (2019-11-06)
    * Fixed config merging helper causing certain user settings to be lost.
    * No longer exposes exception class name for message-less exceptions. Fixes #107